[c++ grpc] Make call data nested structs
