+++ cpp/inc/bond/ext/grpc/detail/unary_call_impl.h
@@ -176,7 +176,22 @@ namespace bond { namespace ext { namespace gRPC { namespace detail {
     template <typename Request, typename Response>
     class unary_call_base
     {
+        using impl_type = unary_call_impl<Request, Response>;
+
     public:
+        unary_call_base() = default;
+
+        explicit unary_call_base(boost::intrusive_ptr<impl_type> impl) noexcept
+            : _impl(std::move(impl))
+        {
+            BOOST_ASSERT(_impl);
+        }
+
+        explicit operator bool() const noexcept
+        {
+            return static_cast<bool>(_impl);
+        }
+
         void swap(unary_call_base& rhs) noexcept
         {
             using std::swap;
@@ -232,22 +247,6 @@ namespace bond { namespace ext { namespace gRPC { namespace detail {
         void FinishWithError(const grpc::Status& status)
         {
             impl().FinishWithError(status);
-        }
-
-    protected:
-        using impl_type = unary_call_impl<Request, Response>;
-
-        unary_call_base() = default;
-
-        explicit unary_call_base(boost::intrusive_ptr<impl_type> impl) noexcept
-            : _impl(std::move(impl))
-        {
-            BOOST_ASSERT(_impl);
-        }
-
-        explicit operator bool() const noexcept
-        {
-            return static_cast<bool>(_impl);
         }
 
     private:
