+++ cpp/inc/bond/ext/grpc/shared_unary_call.h
@@ -5,8 +5,8 @@
 
 #include <bond/core/config.h>
 
-#include <bond/ext/grpc/detail/unary_call_impl.h>
-#include <bond/ext/grpc/unary_call.h>
+#include "detail/unary_call_impl.h"
+#include "unary_call.h"
 
 #include <utility>
