+++ cpp/inc/bond/ext/grpc/unary_call.h
@@ -5,20 +5,13 @@
 
 #include <bond/core/config.h>
 
-#include <bond/ext/grpc/detail/unary_call_impl.h>
+#include "detail/unary_call_impl.h"
 
 #include <boost/smart_ptr/intrusive_ptr.hpp>
 
 #include <utility>
 
 namespace bond { namespace ext { namespace gRPC {
-
-namespace detail {
-
-    template <typename Request, typename Response>
-    class service_unary_call_data;
-
-} // namespace detail
 
 template <typename Request, typename Response>
 class shared_unary_call;
@@ -35,40 +28,21 @@ class shared_unary_call;
 /// @note This class can only be moved. If shared ownership semantics are
 /// needed, convert this to a \ref shared_unary_call with the \ref share
 /// function.
-///
-/// @warning A moved-from unary_call is only valid for destruction or
-/// assignment.
 template <typename Request, typename Response>
 class unary_call final : public detail::unary_call_base<Request, Response>
 {
     using base_type = typename unary_call::unary_call_base;
 
 public:
-    /// @brief The default constructor of unary_call is not callable: the
-    /// Bond-over-gRPC infrastructure will create unary_call instances as
-    /// needed.
-    ///
-    /// Once created, a unary_call instance can only be moved to another
-    /// unary_call or converted to a \ref shared_unary_call (see the \ref
-    /// share function).
-    unary_call() = delete;
+    using base_type::base_type;
 
-    // unary_call is move-only
     unary_call(const unary_call&) = delete;
     unary_call& operator=(const unary_call&) = delete;
 
-    /// @brief Constructs a unary_call by moving from another unary_call.
-    ///
-    /// The moved-from instance is valid only for destruction and
-    /// assignment.
     unary_call(unary_call&& other) = default;
-
-    /// @brief Assign to this unary_call by moving from another unary_call.
-    ///
-    /// The moved-from instance is valid only for destruction and
-    /// assignment.
     unary_call& operator=(unary_call&& rhs) = default;
 
+    using base_type::operator bool;
     using base_type::swap;
 
     /// @brief Creates a \ref shared_unary_call from this instance.
@@ -80,12 +54,6 @@ public:
     {
         return shared_unary_call<Request, Response>{ std::move(*this) };
     }
-
-private:
-    template <typename OtherRequest, typename OtherResponse>
-    friend class detail::service_unary_call_data;
-
-    using base_type::base_type;
 };
 
 template <typename Request, typename Response>
